FastAPI/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyJWT==2.9.0
pydantic==2.9.2
mysql-connector-python==8.4.0
